+++ resources/views/SupportTicketStatus.blade.php
@@ -15,120 +15,42 @@
     <div class="main-block">
         <div class="form-image" style="background-image: url('{{ asset('images/3.jpg') }}');"></div>
 
-        @if (isset($availabledates))
-            <form method="POST" action="/Register/Complete">
-                <button onclick="location.href='/'" type="button"
-                    class="standard-button">{!! __('← На главную') !!}</button>
-                <br><br><br><br>
+        <form method="POST" action="/SupportTicketStatus">
+            <button onclick="location.href='/'" type="button" class="standard-button">{!! __('← На главную') !!}</button>
+            <br><br><br><br>
 
-                @csrf
+            @csrf
 
-                <label>&nbsp{!! __('Выберите удобную дату, аудиторию и время:') !!}</label>
-                <br>
-                <br>
-                <select id="date" name="chosenDate" required>
-                    <option value="" disabled selected>{!! __('Дата') !!}</option>
-                    @foreach ($availabledates as $record)
-                        <option value="{{ $record->date }}">{{ $record->date }}
-                            @if ($record->isOnline == true)
-                                ({!! __('Онлайн') !!})
-                            @else
-                                ({!! __('Офлайн, c ') !!} {{ $record->startHour }} {!! __('до') !!}
-                                {{ $record->endHour }})
-                            @endif
-                        </option>
-                    @endforeach
-                </select>
-                <select id="room" name="roomID" required>
-                    <option value="" disabled selected>{!! __('Аудитория') !!}</option>
-                </select>
-                <select id="startHour" name="startHour" required>
-                    <option value="" disabled selected>{!! __('Время') !!}</option>
-                </select>
-                <input type="hidden" name="mail" value="{{ $mail }}">
-                <input type="hidden" name="requestID" value="{{ $requestID }}">
+            @error('email')
+                <p>{{ $message }}</p>
+            @enderror
+            <input id="email" type="mail" name="email" placeholder=" {!! __('Почта, указанная в заявке') !!}" required />
 
-                <br><br><br>
-                <div style="text-align:center">
-                    <button type="submit" class="standard-button-long">{!! __('Выбрать') !!}</button>
-                </div>
-            </form>
-        @else
-            <form method="POST" action="/Register/Date">
-                <button onclick="location.href='/'" type="button"
-                    class="standard-button">{!! __('← На главную') !!}</button>
-                <br><br><br><br>
 
-                @csrf
+            @error('supportTicketID')
+                <p>{{ $message }}</p>
+            @enderror
+            <input id="supportTicketID" type="number" name="supportTicketID" placeholder=" {!! __('Номер заявки') !!}"
+                required />
 
-                <input type="email" name="mail" value="" placeholder=" {!! __('Почта, указанная в заявке') !!}" required />
-                <input type="number" name="requestID" value="" placeholder=" {!! __('Номер заявки') !!}"
-                    required />
-
-                <br><br><br>
-                <div style="text-align:center">
-                    <button type="submit" class="standard-button-long">{!! __('Выбрать') !!}</button>
-                </div>
-            </form>
-        @endif
+            <br><br><br>
+            <div style="text-align:center">
+                <button type="submit" class="standard-button-long">{!! __('Выбрать') !!}</button>
+            </div>
+        </form>
     </div>
 </body>
 
 <script src="https://code.jquery.com/jquery-3.7.0.min.js"
     integrity="sha256-2Pmvv0kuTBOenSvLm6bvfBSSHrUJ+3A7x6P5Ebd07/g=" crossorigin="anonymous"></script>
+
+<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.inputmask/5.0.8/jquery.inputmask.min.js"
+    integrity="sha512-efAcjYoYT0sXxQRtxGY37CKYmqsFVOIwMApaEbrxJr4RwqVVGw8o+Lfh/+59TU07+suZn1BWq4fDl5fdgyCNkw=="
+    crossorigin="anonymous" referrerpolicy="no-referrer"></script>
 <script>
-    var chosenDate;
-
-    $('#date').change(function() {
-        chosenDate = $(this).val();
-
-        $('#room').empty();
-        $('#room').append(
-            '<option value="" disabled selected>{!! __('Аудитория') !!}</option>'
-        );
-        $('#startHour').empty();
-        $('#startHour').append(
-            '<option value="" disabled selected>{!! __('Время') !!}</option>'
-        );
-
-        $.ajax({
-            type: 'GET',
-            url: '/Register/Room',
-            data: {
-                chosenDate
-            },
-            success: function(rooms) {
-                if (rooms.isOnline) {
-                    $('#room').empty();
-                    $('#room').append(
-                        '<option value="isOnline" selected>{!! __('Онлайн') !!}</option>'
-                    );
-                    $('#startHour').empty();
-                    $('#startHour').append(
-                        '<option value="isOnline" selected>{!! __('Онлайн') !!}</option>'
-                    );
-                } else {
-                    $('#room').append(rooms);
-                }
-            }
-        });
-    });
-
-    $('#room').change(function() {
-        $('#startHour').children().not(':first').remove();
-        roomID = $(this).val();
-
-        $.ajax({
-            type: 'GET',
-            url: '/Register/Hour',
-            data: {
-                chosenDate,
-                roomID
-            },
-            success: function(hours) {
-                $('#startHour').append(hours);
-            }
-        });
+    // Charater "a" with \\, because it is special character
+    $('#email').inputmask({
+        mask: "*{1,20}@\\alm\\au.edu.kz",
     });
 </script>